Group.InRange
